requirements.txt
# LangChain (0.3.x compatible stack)
langchain==0.3.20
langchain-core==0.3.83
langchain-community==0.3.18
langchain-text-splitters==0.3.8
langchain-huggingface==0.3.1

# Embeddings & Vector DB
sentence-transformers==2.7.0
faiss-cpu==1.8.0

# HuggingFace
torch==2.2.2
transformers==4.40.2

# Utilities
youtube-transcript-api==0.6.2
python-dotenv==1.0.1
